Fix headers should only increase by one
The challenge is, various pages have various headers, some have an h1 and h2, some have h1,h2,h3,h4, etc.

The footer doesn't "know" how many headers have been there before it so we can't adjust that to the correct number.

The proposed fix is that we should change this to a nav section since it's just links to social media and that's navigation.