[a11y] Fixup skip to main content list tag. (#2105)
Lighthouse was showing an error about the skip to main content `li` tag
not being contained in a `ul`. The a11-tools-query gave a bit more
context that the `ul` should not have a `navigation` role (fixing both
the invalid `navigation` role on the `ul` and the `li` missing a
non-role'd parent).

This change uses the same `ul` as is used on Canada.ca and to wrap the
`ul` in a `nav` similar to Canada.ca.