Fixup .File.Dir warning from Hugo.
This PR removes the .File.Dir warning generated by Hugo on startup. It
does this by removing the use of .File.Dir and adding a new key into the
parameters for each of our types of things (partnership, platform-tool
or resource). This key is then used `.Params.productkind` to do the
filtering instead of the `.File.Dir` check.

Issue #2052