Change card headings for better accessibility. (#2322)
This PR changes the headings of the product cards to use h1/h2 instead
of h4/h5. In the latter case, the homepage would have headings h1, h2,
h4, h5 which causes an accessibility warning.

This PR changes the cards to use h1/h2 as the headings are inside a
section and updates the stylings to match. Most of the divs have been
removed as the cards still appear to format correctly without all the
tag nesting.

Fixes #2288