Change heading structure for partnership list.
This PR changes the partnership include to not use h1/h2 anymore.
Instead div's are used which are styled with classes `.h1` and `.h2`.
The top level div is marked as a `role=heading` but the others are not
as they don't really work like headings in the card.

Issue #2288