Revert "Change card headings for better accessibility. (#2322)"
This reverts commit 0b9268ab70b91a3e1aa8bfff9afb3a1199fd2a3d.

When looking at the changes it was decided it was better to skip a
header level on the main page then to have duplicate header levels which
happened with the fix. We were unable to get the current system to take
the header level as a template parameter.

This will go back to the backlog and we'll probably need to either
duplicate the list template, or change the site so the header levels are
the same when this template gets included.

Issue #2288