Chore/meet the team fix (#2403)
Fixed up the meet the team page. In the previous PR, i changed the team structure to organize it based on roles but this caused some issues with the PR Bot when submitting new content. Instead of having to change both the PR bot and the CMS repo, I redid the meet the team page by reverting the team structure to what it originally was followed by other changes to group the team under their specific heading.