feat: add an aria-label to "View this product"
Adds an aria label to the View this product links that lets screen
readers know what page they will go to.

To simplify translation I'm simply using the title of the card so we
don't have to worry about handling products that start with vowels and
articles if we want to use View the.

This may need some content design thinking to determine the best text we
want to use for those folks viewing the site with screenreaders.