fix: adds a label to an svg object
This fixes a serious a11y issue. The fix was to copy the aria label from
the fip.html page. I didn't replace the FIP in the engagement html
because I didn't know why it wasn't using it in the first place and
wanted to make sure the change was minimal.