Revert "Redesign/job posting redesign (#2394)"
This reverts commit b2bbb9a3cfd94c7c2e5fc9a784398496240125ce.

There was a bad merge of the CSS files which is breaking the styles on
all of the product cards along with the blog post headers and other
unknown places. This needs the CSS merge redone, the formatting changes
should be done in a prior pr to make reviewing easier.